move Instrument to editorStore dir
