+++ src/modules/editorStore/Instrument.ts
